remove cli subcrate, and move tests into stop_areas
