+++ Cargo.toml
@@ -8,7 +8,6 @@ edition = "2018"
 
 [workspace]
 members = [
-	"cli",
 	"server",
 	"launch",
 	"random",
