--- cli/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "loki_cli"
-version = "0.1.0"
-authors = ["Pascal Benchimol <pascal.benchimol@kisio.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-launch = { path = "../launch"}
-structopt = "0.3"
-serde = "1.0"
-serde_json = "1.0"
-chrono = "0.4"
-failure = "0.1"
-slog = "2.5"
-slog-term = "2.4"
-slog-scope = "4.1"
-slog-stdlog = "4.0"
-slog-async = "2.3"
-slog-envlogger = "2.1"
-rand = "0.7"
-rand_chacha = "0.2.2"
-log = "0.4"
-# log = { version = "0.4", features = ["max_level_debug", "release_max_level_warn"] }
