+++ stop_areas/src/main.rs
@@ -34,30 +34,9 @@
 // https://groups.google.com/d/forum/navitia
 // www.navitia.io
 
-use launch::{config, loki::{DailyData, LoadsDailyData, LoadsPeriodicData, PeriodicData}, solver};
-use launch::loki;
-
-use loki::log;
-
-use loki::{log::trace,transit_model::Model,
-};
-
-use loki::traits;
-
-
-use std::{ fs::File, io::BufReader, time::SystemTime};
-
-
-use failure::{bail, Error};
-
-
-
-use serde::{Serialize, Deserialize};
-use structopt::StructOpt;
-
 fn main() {
     let _log_guard = launch::logger::init_logger();
-    if let Err(err) = run() {
+    if let Err(err) = stop_areas::run() {
         for cause in err.iter_chain() {
             eprintln!("{}", cause);
         }
@@ -66,199 +45,4 @@ fn main() {
 }
 
 
-#[derive(StructOpt)]
-#[structopt(
-    name = "loki_stop_areas",
-    about = "Perform a public transport request between two stop areas.",
-    rename_all = "snake_case"
-)]
-pub enum Options {
-    /// Create a config file from cli arguments
-    CreateConfig(ConfigCreator),
-    /// Launch from a config file
-    ConfigFile(ConfigFile),
-    /// Launch from cli arguments
-    Launch(Config),
-}
-
-#[derive(StructOpt)]
-#[structopt(
-    rename_all = "snake_case"
-)]
-pub struct ConfigCreator {
-    #[structopt(flatten)]
-    pub config: Config,
-
-}
-
-#[derive(StructOpt)]
-pub struct ConfigFile {
-    /// path to the json config file
-    #[structopt(parse(from_os_str))]
-    file: std::path::PathBuf,
-}
-#[derive(Serialize, Deserialize)]
-#[derive(StructOpt)]
-#[structopt(
-    rename_all = "snake_case"
-)]
-pub struct Config {
-    #[serde(flatten)]
-    #[structopt(flatten)]
-    pub launch_params : config::LaunchParams,
-
-    #[serde(flatten)]
-    #[structopt(flatten)]
-    pub request_params: config::RequestParams,
-
-    /// Departure datetime of the query, formatted like 20190628T163215
-    /// If none is given, all queries will be made at 08:00:00 on the first
-    /// valid day of the dataset
-    #[structopt(long)]
-    pub departure_datetime: Option<String>,
-
-    /// Which comparator to use for the request
-    /// "basic" or "loads"
-    #[serde(default)]
-    #[structopt(long, default_value)]
-    pub comparator_type: config::ComparatorType,
-
-    /// name of the start stop_area
-    #[structopt(long)]
-    pub start: String,
-
-    /// name of the end stop_area
-    #[structopt(long)]
-    pub end: String,
-}
-
-
-
-pub fn run() -> Result<(), Error> {
-    let options = Options::from_args(); 
-    match options {
-        Options::ConfigFile(config_file) => {
-            let config = read_config(&config_file)?;
-            launch(config)?;
-            Ok(())
-        },
-        Options::CreateConfig(config_creator) => {
-            let json_string = serde_json::to_string_pretty(&config_creator.config)?;
-
-            println!("{}", json_string);
-
-            Ok(())
-        }
-        Options::Launch(config) => {
-            launch(config)?;
-            Ok(())
-        }
-    }
-
-}
-
-
-
-
-
-pub fn read_config(config_file : & ConfigFile) -> Result<Config, Error> {
-    let file = match File::open(&config_file.file) {
-        Ok(file) => file,
-        Err(e) => {
-            bail!("Error opening config file {:?} : {}", &config_file.file, e)
-        }
-    };
-    let reader = BufReader::new(file);
-    let config : Config = serde_json::from_reader(reader).map_err(|err| {
-        failure::format_err!("Could not read config file {:?} : {}", config_file.file, err)
-    })?;
-    Ok(config)
-}
-
-pub fn launch(config :  Config) -> Result<(Model, Vec<loki::Response>), Error> {
-
-    match config.launch_params.data_implem {
-        config::DataImplem::Periodic => {
-            config_launch::<PeriodicData>(config)
-        }
-        config::DataImplem::Daily => {
-            config_launch::<DailyData>(config)
-        }
-        config::DataImplem::LoadsPeriodic => {
-            config_launch::<LoadsPeriodicData>(config)
-        }
-        config::DataImplem::LoadsDaily => {
-            config_launch::<LoadsDailyData>(config)
-        }
-    }
-}
-
-fn config_launch<Data>(config: Config) -> Result<(Model, Vec<loki::Response>), Error>
-where
-    Data: traits::DataWithIters,
-{
-    let (data, model) = launch::read(
-        &config.launch_params
-    )?;
-    let result = match config.launch_params.criteria_implem {
-        config::CriteriaImplem::Basic => 
-        {
-            build_engine_and_solve::<Data,solver::BasicCriteriaSolver<Data>>
-                (&model, &data, &config)
-        },
-        config::CriteriaImplem::Loads => 
-        {
-            build_engine_and_solve::<Data, solver::LoadsCriteriaSolver<Data>>
-                (&model, &data, &config)
-        },
-    };
-
-    result.map(|responses| (model, responses))
-}
-
-fn build_engine_and_solve< Data, Solver>(
-    model: &Model,
-    data: & Data,
-    config: &Config,
-) -> Result<Vec<loki::Response>, Error>
-where
-    Data: traits::DataWithIters,
-    Solver: solver::Solver<Data>,
-{
-    let mut solver = Solver::new(data.nb_of_stops(), data.nb_of_missions());
-
-    let departure_datetime = match &config.departure_datetime {
-        Some(string_datetime) => launch::datetime::parse_datetime(&string_datetime)?,
-        None => {
-            let naive_date = data.calendar().first_date();
-            naive_date.and_hms(8, 0, 0)
-        }
-    };
-
-    let compute_timer = SystemTime::now();
-
-    let start_stop_area_uri = &config.start;
-    let end_stop_area_uri = &config.end;
-
-    let request_input = launch::stop_areas::make_query_stop_areas(model, &departure_datetime, start_stop_area_uri, end_stop_area_uri, &config.request_params)?;
-    let solve_result = solver.solve_request(data, model, request_input, &config.comparator_type);
-
-
-    let duration = compute_timer.elapsed().unwrap().as_millis();
-    log::info!("Duration : {} ms", duration as f64);
-
-    match &solve_result {
-        Err(err) => {
-            log::error!("Error while solving request : {}", err);
-        }
-        Ok(responses) => {
-            for response in responses.iter() {
-                trace!("{}", response.print(model)?);
-            }
-        }
-    }
-
-    let responses = solve_result?;
-    Ok(responses)
-}
