+++ stop_areas/tests/one_line/companies.txt
