+++ stop_areas/tests/one_line/lines.txt
