+++ stop_areas/tests/one_line/physical_modes.txt
