+++ stop_areas/tests/one_line/stops.txt
