+++ stop_areas/tests/one_line/trips.txt
